Implement `DiffEqBase.anyeltypedual` for GraphDynamics types
The PR https://github.com/Neuroblox/GraphDynamics.jl/pull/32 somehow broke a
couple things in the Neuroblox.jl docs building pipeline due to how it
interacted with the GraphDynamics tutorial. This change (which is good
to do anyways) should fix that breakage. It basically just helps the
differential equation solvers tell if anything in the problem is a
dual number.

I also did a little cleanup of unwanted stuff, and implemented a
couple SymbolicIndexingInterface methods for `Num`s with observed variables.